+++ apps/laboratory/src/pages/library/solana-email.tsx
@@ -1,0 +1,46 @@
+import { createWeb3Modal, defaultSolanaConfig } from '@web3modal/solana/react'
+
+import { ThemeStore } from '../../utils/StoreUtil'
+import { solana, solanaDevnet, solanaTestnet } from '../../utils/ChainsUtil'
+import { AppKitButtons } from '../../components/AppKitButtons'
+import { ConstantsUtil } from '../../utils/ConstantsUtil'
+import { SolanaTests } from '../../components/Solana/SolanaTests'
+import { SolflareWalletAdapter } from '@solana/wallet-adapter-wallets'
+
+const chains = [solana, solanaTestnet, solanaDevnet]
+
+export const solanaConfig = defaultSolanaConfig({
+  chains,
+  projectId: ConstantsUtil.ProjectId,
+  metadata: ConstantsUtil.Metadata,
+  auth: {
+    email: true,
+    socials: ['google', 'x', 'discord', 'farcaster', 'github', 'apple', 'facebook'],
+    walletFeatures: true,
+    showWallets: true
+  }
+})
+
+const modal = createWeb3Modal({
+  solanaConfig,
+  projectId: ConstantsUtil.ProjectId,
+  metadata: ConstantsUtil.Metadata,
+  chains,
+  enableAnalytics: false,
+  termsConditionsUrl: 'https://walletconnect.com/terms',
+  privacyPolicyUrl: 'https://walletconnect.com/privacy',
+  customWallets: ConstantsUtil.CustomWallets,
+  enableSwaps: false,
+  wallets: [new SolflareWalletAdapter()]
+})
+
+ThemeStore.setModal(modal)
+
+export default function Solana() {
+  return (
+    <>
+      <AppKitButtons />
+      <SolanaTests />
+    </>
+  )
+}