+++ apps/laboratory/tests/shared/pages/ModalPage.ts
@@ -254,11 +254,14 @@ export class ModalPage {
     ).toBeVisible({
       timeout: 10000
     })
-    await this.page.waitForTimeout(2000)
+    await this.page.waitForTimeout(500)
   }
 
   async clickSignatureRequestButton(name: string) {
+    const signatureHeader = this.page.getByText('Approve Transaction')
     await this.page.frameLocator('#w3m-iframe').getByRole('button', { name, exact: true }).click()
+    await expect(signatureHeader, 'Signature request should be closed').not.toBeVisible()
+    await this.page.waitForTimeout(300)
   }
 
   async approveSign() {
@@ -317,6 +320,9 @@ export class ModalPage {
   }
 
   async openAccount() {
+    expect(this.page.getByTestId('w3m-modal-card')).not.toBeVisible()
+    expect(this.page.getByTestId('w3m-modal-overlay')).not.toBeVisible()
+    this.page.waitForTimeout(300)
     await this.page.getByTestId('account-button').click()
   }
 
@@ -372,6 +378,15 @@ export class ModalPage {
     await this.enterOTP(otp, headerTitle)
   }
 
+  async switchNetworkWithNetworkButton(networkName: string) {
+    const networkButton = this.page.getByTestId('w3m-network-button')
+    await networkButton.click()
+
+    const networkToSwitchButton = this.page.getByTestId(`w3m-network-switch-${networkName}`)
+    await networkToSwitchButton.click()
+    await networkToSwitchButton.waitFor({ state: 'hidden' })
+  }
+
   async openModal() {
     await this.page.getByTestId('account-button').click()
   }
@@ -404,15 +419,6 @@ export class ModalPage {
 
     await sendCallsInput.fill(batchCallId)
     await sendCallsButton.click()
-  }
-
-  async switchNetworkWithNetworkButton(networkName: string) {
-    const networkButton = this.page.getByTestId('w3m-network-button')
-    await networkButton.click()
-
-    const networkToSwitchButton = this.page.getByTestId(`w3m-network-switch-${networkName}`)
-    await networkToSwitchButton.click()
-    await networkToSwitchButton.waitFor({ state: 'hidden' })
   }
 
   async switchAccount() {