+++ apps/laboratory/tests/smart-account.spec.ts
@@ -46,11 +46,10 @@ smartAccountTest.afterAll(async () => {
 })
 
 // -- Tests --------------------------------------------------------------------
-smartAccountTest('it should use a Smart Account', async () => {
+smartAccountTest('it should use a smart account', async () => {
   await validator.expectConnected()
   await page.openAccount()
   await validator.expectActivateSmartAccountPromoVisible(false)
-
   await page.openProfileView()
   await page.openSettings()
   await validator.expectChangePreferredAccountToShow(EOA)
@@ -69,27 +68,13 @@ smartAccountTest('it should sign with smart account 6492 signature', async () =>
   await validator.expectValidSignature(signature, address, chainId)
 })
 
-smartAccountTest('it should switch to a SA enabled network and sign', async () => {
-  const targetChain = 'Sepolia'
-  await page.openAccount()
-  await page.openProfileView()
-  await page.openSettings()
-  await page.switchNetwork(targetChain)
-  await validator.expectSwitchedNetwork(targetChain)
-  await page.closeModal()
-  await page.sign()
-  await page.approveSign()
-  await validator.expectAcceptedSign()
-})
-
 smartAccountTest('it should switch to a not enabled network and sign with EOA', async () => {
   const targetChain = 'Ethereum'
-  await page.openAccount()
-  await page.openProfileView()
-  await page.openSettings()
+  await page.goToSettings()
   await page.switchNetwork(targetChain)
-  await validator.expectSwitchedNetwork(targetChain)
-  // Shouldn't show the toggle on a non enabled network
+
+  await page.page.waitForTimeout(2000)
+  await page.goToSettings()
   await validator.expectTogglePreferredTypeVisible(false)
   await page.closeModal()
 
@@ -99,16 +84,14 @@ smartAccountTest('it should switch to a not enabled network and sign with EOA', 
 })
 
 smartAccountTest('it should switch to smart account and sign', async () => {
-  await page.openAccount()
-  await page.openProfileView()
-  await page.openSettings()
-
-  await page.switchNetwork('Polygon')
-  await validator.expectSwitchedNetwork('Polygon')
+  const targetChain = 'Polygon'
+  await page.goToSettings()
+  await page.switchNetwork(targetChain)
+  await validator.expectSwitchedNetwork(targetChain)
 
   await page.togglePreferredAccountType()
+  await page.goToSettings()
   await validator.expectChangePreferredAccountToShow(EOA)
-
   await page.closeModal()
 
   await page.sign()
@@ -123,30 +106,19 @@ smartAccountTest('it should switch to smart account and sign', async () => {
 })
 
 smartAccountTest('it should switch to eoa and sign', async () => {
-  await page.openAccount()
-  await page.openProfileView()
-  await page.openSettings()
-
+  await page.goToSettings()
   await page.togglePreferredAccountType()
+  await page.goToSettings()
   await validator.expectChangePreferredAccountToShow(SMART_ACCOUNT)
-
   await page.closeModal()
 
   await page.sign()
   await page.approveSign()
   await validator.expectAcceptedSign()
-
-  const signature = await page.getSignature()
-  const address = await page.getAddress()
-  const chainId = await page.getChainId()
-
-  await validator.expectValidSignature(signature, address, chainId)
 })
 
 smartAccountTest('it should disconnect correctly', async () => {
-  await page.openAccount()
-  await page.openProfileView()
-  await page.openSettings()
+  await page.goToSettings()
   await page.disconnect()
   await validator.expectDisconnected()
 })