+++ packages/scaffold-ui/src/views/w3m-connecting-external-view/index.ts
@@ -46,7 +46,7 @@ export class W3mConnectingExternalView extends W3mConnectingWidget {
          * And if there is an error, this condition will be skipped and the connection will be triggered as usual because we have `Try again` button in this view which is a user interaction as well.
          */
         if (this.connector.id !== ConstantsUtil.COINBASE_SDK_CONNECTOR_ID || !this.error) {
-          await ConnectionController.connectExternal(this.connector)
+          await ConnectionController.connectExternal(this.connector, this.connector.chain)
 
           if (OptionsController.state.isSiweEnabled) {
             RouterController.push('ConnectingSiwe')